refactor: 노드 설정 세션 제거
