+++ src/features/configure-node/model/index.ts
@@ -1,2 +1,6 @@
 export * from "./panelRegistry";
+export * from "./setup-types";
+export * from "./sink-node-draft";
+export * from "./source-node-draft";
+export * from "./source-target-schema";
 export * from "./types";
