+++ src/features/configure-node/ui/panels/AuthPrompt.tsx
