+++ src/features/configure-node/ui/panels/index.ts
@@ -2,3 +2,4 @@ export * from "./CommunicationPanel";
 export * from "./GenericNodePanel";
 export * from "./NodePanelShell";
 export * from "./SinkNodePanel";
+export * from "./SourceNodePanel";
