--- src/features/node-setup/index.ts
@@ -1,2 +1,0 @@
-export * from "./model";
-export * from "./ui";
