--- src/features/node-setup/model/index.ts
@@ -1,4 +1,0 @@
-export * from "./sink-node-draft";
-export * from "./source-node-draft";
-export * from "./source-target-schema";
-export * from "./types";
