--- src/features/node-setup/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./NodeSetupOverlay";
