+++ src/features/workflow-editor/model/workflowStore.ts
@@ -26,14 +26,6 @@ export type PlaceholderInfo = {
 
 export type ActivePanelMode = "view" | "edit";
 
-export type NodeSetupRole = "start" | "middle" | "end";
-
-export type ActiveNodeSetupSession = {
-  mode: "edit";
-  nodeId: string;
-  role: NodeSetupRole;
-} | null;
-
 type NodePosition = {
   x: number;
   y: number;
@@ -52,7 +44,6 @@ interface WorkflowEditorState {
   nodeStatuses: WorkflowNodeStatusMap;
   activePanelNodeId: string | null;
   activePanelMode: ActivePanelMode;
-  activeNodeSetupSession: ActiveNodeSetupSession;
   startNodeId: string | null;
   endNodeId: string | null;
   activePlaceholder: PlaceholderInfo | null;
@@ -75,8 +66,6 @@ interface WorkflowEditorActions {
   replaceNodeConfig: (id: string, config: FlowNodeData["config"]) => void;
   openPanel: (nodeId: string, options?: { mode?: ActivePanelMode }) => void;
   setActivePanelMode: (mode: ActivePanelMode) => void;
-  openNodeSetup: (session: Exclude<ActiveNodeSetupSession, null>) => void;
-  closeNodeSetup: () => void;
   closePanel: () => void;
   setWorkflowMeta: (id: string, name: string) => void;
   hydrateWorkflow: (payload: WorkflowHydratedState) => void;
@@ -85,7 +74,6 @@ interface WorkflowEditorActions {
     options?: {
       preserveActivePanelNodeId?: boolean;
       preserveActivePlaceholder?: boolean;
-      preserveActiveNodeSetupSession?: boolean;
       preserveDirty?: boolean;
     },
   ) => void;
@@ -108,7 +96,6 @@ const initialState: WorkflowEditorState = {
   nodeStatuses: {},
   activePanelNodeId: null,
   activePanelMode: "view",
-  activeNodeSetupSession: null,
   startNodeId: null,
   endNodeId: null,
   activePlaceholder: null,
@@ -249,33 +236,17 @@ export const useWorkflowStore = create<
       set((state) => {
         state.activePanelNodeId = nodeId;
         state.activePanelMode = options?.mode ?? "view";
-        state.activeNodeSetupSession = null;
       }),
 
     setActivePanelMode: (mode) =>
       set((state) => {
         state.activePanelMode = mode;
-        state.activeNodeSetupSession = null;
-      }),
-
-    openNodeSetup: (session) =>
-      set((state) => {
-        state.activeNodeSetupSession = session;
-        state.activePanelNodeId = session.nodeId;
-        state.activePanelMode = "view";
-        state.activePlaceholder = null;
-      }),
-
-    closeNodeSetup: () =>
-      set((state) => {
-        state.activeNodeSetupSession = null;
       }),
 
     closePanel: () =>
       set((state) => {
         state.activePanelNodeId = null;
         state.activePanelMode = "view";
-        state.activeNodeSetupSession = null;
       }),
 
     setStartNodeId: (id) =>
@@ -297,9 +268,6 @@ export const useWorkflowStore = create<
     setActivePlaceholder: (placeholder) =>
       set((state) => {
         state.activePlaceholder = placeholder;
-        if (placeholder) {
-          state.activeNodeSetupSession = null;
-        }
       }),
 
     applyLayoutPositions: (updates) =>
@@ -354,7 +322,6 @@ export const useWorkflowStore = create<
         state.endNodeId = payload.endNodeId;
         state.activePanelNodeId = null;
         state.activePanelMode = "view";
-        state.activeNodeSetupSession = null;
         state.activePlaceholder = null;
         state.unsavedNodePositions = {};
         state.isDirty = false;
@@ -367,8 +334,6 @@ export const useWorkflowStore = create<
           options?.preserveActivePanelNodeId ?? true;
         const preserveActivePlaceholder =
           options?.preserveActivePlaceholder ?? true;
-        const preserveActiveNodeSetupSession =
-          options?.preserveActiveNodeSetupSession ?? true;
         const preserveDirty = options?.preserveDirty ?? true;
         const nextNodeIds = new Set(payload.nodes.map((node) => node.id));
         const preservedUnsavedNodePositions = Object.fromEntries(
@@ -396,11 +361,6 @@ export const useWorkflowStore = create<
         if (!state.activePanelNodeId) {
           state.activePanelMode = "view";
         }
-        state.activeNodeSetupSession =
-          preserveActiveNodeSetupSession &&
-          hasNode(payload.nodes, state.activeNodeSetupSession?.nodeId ?? null)
-            ? state.activeNodeSetupSession
-            : null;
         if (!preserveActivePlaceholder) {
           state.activePlaceholder = null;
         }
