+++ src/widgets/output-panel/ui/OutputPanel.tsx
@@ -113,9 +113,6 @@ export const OutputPanel = ({ wizardController }: Props) => {
   const activePlaceholder = useWorkflowStore(
     (state) => state.activePlaceholder,
   );
-  const activeNodeSetupSession = useWorkflowStore(
-    (state) => state.activeNodeSetupSession,
-  );
   const workflowId = useWorkflowStore((state) => state.workflowId);
   const nodeStatuses = useWorkflowStore((state) => state.nodeStatuses);
   const activePanelMode = useWorkflowStore((state) => state.activePanelMode);
@@ -132,17 +129,15 @@ export const OutputPanel = ({ wizardController }: Props) => {
   const setActivePanelMode = useWorkflowStore(
     (state) => state.setActivePanelMode,
   );
-  const openNodeSetup = useWorkflowStore((state) => state.openNodeSetup);
   const layout = useDualPanelLayout();
-  const isOpen =
-    Boolean(activePanelNodeId) &&
-    activePlaceholder === null &&
-    activeNodeSetupSession === null;
-
   const activeNode = useMemo(
     () => nodes.find((node) => node.id === activePanelNodeId) ?? null,
     [activePanelNodeId, nodes],
   );
+  const isStartNode = Boolean(activeNode && activeNode.id === startNodeId);
+  const isEndNode = Boolean(activeNode && activeNode.id === endNodeId);
+  const isOpen =
+    Boolean(activePanelNodeId) && activePlaceholder === null && !isStartNode;
 
   const activeNodeStatus = activePanelNodeId
     ? (nodeStatuses[activePanelNodeId] ?? null)
@@ -154,7 +149,13 @@ export const OutputPanel = ({ wizardController }: Props) => {
       (activeNodeStatus.missingFields?.length ?? 0) > 0),
   );
   const isEditMode = activePanelMode === "edit" && canEditNodes;
-  const isDetailMode = !isEditMode && isMiddleWizardCompleted(activeNode);
+  const isEndEditMode = isEditMode && isEndNode;
+  const isEndViewMode = !isEditMode && isEndNode;
+  const isDetailMode =
+    !isEditMode &&
+    !isStartNode &&
+    !isEndNode &&
+    isMiddleWizardCompleted(activeNode);
   const nodeDataPanel = useNodeDataPanelModel({
     panelKind: "output",
     workflowId: workflowId || undefined,
@@ -163,8 +164,6 @@ export const OutputPanel = ({ wizardController }: Props) => {
     isWorkflowDirty: isDirty,
   });
   const activeMeta = activeNode ? NODE_REGISTRY[activeNode.data.type] : null;
-  const isStartNode = Boolean(activeNode && activeNode.id === startNodeId);
-  const isEndNode = Boolean(activeNode && activeNode.id === endNodeId);
   const outputDataLabel = nodeDataPanel.staticOutputLabel ?? "출력 데이터";
   const hasPreviewData = !isEmptyPanelData(nodeDataPanel.dataToDisplay);
   const shouldShowSchemaPreview =
@@ -334,6 +333,93 @@ export const OutputPanel = ({ wizardController }: Props) => {
             </Text>
           ) : null}
         </>
+      ) : isEndEditMode && activeNode ? (
+        <>
+          <Box
+            display="flex"
+            alignItems="center"
+            justifyContent="space-between"
+            px={3}
+          >
+            <Text fontSize="xl" fontWeight="medium" letterSpacing="-0.4px">
+              설정
+            </Text>
+            <Box cursor="pointer" onClick={handleClose}>
+              <Icon as={MdCancel} boxSize={6} color="gray.600" />
+            </Box>
+          </Box>
+
+          <Box flex={1} overflow="auto" p={3}>
+            <PanelRenderer
+              readOnly={!canEditNodes}
+              onCancel={() => setActivePanelMode("view")}
+              onComplete={() => setActivePanelMode("view")}
+            />
+          </Box>
+        </>
+      ) : isEndViewMode && activeNode && activeMeta ? (
+        <>
+          <Box
+            display="flex"
+            alignItems="center"
+            justifyContent="space-between"
+            px={3}
+          >
+            <Box display="flex" gap={2} alignItems="center">
+              <Icon
+                as={activeMeta.iconComponent}
+                boxSize={6}
+                color={activeMeta.color}
+              />
+              <Text fontSize="xl" fontWeight="medium" letterSpacing="-0.4px">
+                보낼 곳
+              </Text>
+            </Box>
+            <Box cursor="pointer" onClick={handleClose}>
+              <Icon as={MdCancel} boxSize={6} color="gray.600" />
+            </Box>
+          </Box>
+
+          <VStack align="stretch" flex={1} overflow="auto" p={3} gap={6}>
+            <Box>
+              <Text fontSize="lg" fontWeight="bold" mb={2}>
+                {activeMeta.label}
+              </Text>
+              <Text fontSize="sm" color="text.secondary">
+                처리 결과를 보낼 대상과 세부 설정을 확인합니다.
+              </Text>
+            </Box>
+
+            {hasConfigIssue ? (
+              <Box
+                bg="orange.50"
+                border="1px solid"
+                borderColor="orange.100"
+                borderRadius="2xl"
+                px={4}
+                py={4}
+              >
+                <Text color="orange.600" fontSize="sm" fontWeight="semibold">
+                  설정 확인 필요
+                </Text>
+                <Text mt={1} color="text.secondary" fontSize="sm">
+                  실행 전에 이 도착 노드의 설정을 다시 확인해 주세요.
+                </Text>
+              </Box>
+            ) : null}
+
+            {canEditNodes ? (
+              <Button
+                alignSelf="flex-start"
+                size="sm"
+                variant="outline"
+                onClick={() => setActivePanelMode("edit")}
+              >
+                설정 수정
+              </Button>
+            ) : null}
+          </VStack>
+        </>
       ) : isDetailMode && activeNode && activeMeta ? (
         <>
           <Box
@@ -467,23 +553,6 @@ export const OutputPanel = ({ wizardController }: Props) => {
           </Box>
 
           <Box flex={1} overflow="auto">
-            {canEditNodes && activeNode && (isStartNode || isEndNode) ? (
-              <Box px={3} pb={4}>
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={() =>
-                    openNodeSetup({
-                      mode: "edit",
-                      nodeId: activeNode.id,
-                      role: isStartNode ? "start" : "end",
-                    })
-                  }
-                >
-                  설정 수정
-                </Button>
-              </Box>
-            ) : null}
             <PanelRenderer readOnly={!canEditNodes} />
           </Box>
         </>
